+++ lib/pod/command/trunk/push.rb
@@ -75,6 +75,7 @@ module Pod
         private
 
         def push_to_trunk
+          spec.attributes_hash[:pushed_with_swift_version] = @swift_version
           response = request_path(:post, "pods?allow_warnings=#{@allow_warnings}",
                                   spec.to_json, auth_headers)
           url = response.headers['location'].first
@@ -119,6 +120,10 @@ module Pod
           unless validator.validated?
             raise Informative, "The spec did not pass validation, due to #{validator.failure_reason}."
           end
+
+          # Let the validator's logic for the swift version
+          # set the value for the trunk JSON uploader
+          @swift_version = validator.used_swift_version
         end
 
         def update_master_repo